+++ src/quacc/recipes/common/phonons.py
@@ -7,7 +7,6 @@ from monty.dev import requires
 
 from quacc import flow, job, subflow
 from quacc.atoms.phonons import atoms_to_phonopy, phonopy_atoms_to_ase_atoms
-from quacc.recipes.common.core import force_job
 from quacc.schemas.phonopy import summarize_phonopy
 
 try:
@@ -16,11 +15,10 @@ except ImportError:
     phonopy = None
 
 if TYPE_CHECKING:
-    from typing import Any
+    from typing import Any, Callable
 
     from ase.atoms import Atoms
-    from ase.calculators.calculator import Calculator
-    from numpy.typing import ArrayLike, NDArray
+    from numpy.typing import ArrayLike
     from phonopy import Phonopy
 
     from quacc.schemas._aliases.phonopy import PhononSchema
@@ -30,7 +28,8 @@ if TYPE_CHECKING:
 @requires(phonopy, "Phonopy must be installed. Run `pip install quacc[phonons]`")
 def phonon_flow(
     atoms: Atoms,
-    calculator: Calculator,
+    static_job: Callable,
+    static_job_kwargs: dict[str, Any] | None = None,
     supercell_matrix: ArrayLike = ((2, 0, 0), (0, 2, 0), (0, 0, 2)),
     atom_disp: float = 0.015,
     t_step: float = 10,
@@ -47,8 +46,10 @@ def phonon_flow(
     ----------
     atoms
         Atoms object with calculator attached.
-    calculator
-        Calculator to use.
+    static_job
+        The static job to calculate the forces.
+    static_job_kwargs
+        The kwargs for `static_job`.
     supercell_matrix
         Supercell matrix to use. Defaults to 2x2x2 supercell.
     atom_disp
@@ -67,20 +68,28 @@ def phonon_flow(
     PhononSchema
         Dictionary of results from [quacc.schemas.phonopy.summarize_phonopy][]
     """
-    fields_to_store = fields_to_store or {}
+
+    static_job_kwargs = static_job_kwargs or {}
 
     @subflow
-    def _force_job_distributed(supercells: list[Atoms]) -> list[NDArray]:
+    def _phonopy_forces_subflow(atoms: Atoms) -> list[dict]:
+        phonon = atoms_to_phonopy(atoms, supercell_matrix, atom_disp)
+        supercells = [
+            phonopy_atoms_to_ase_atoms(s) for s in phonon.supercells_with_displacements
+        ]
         return [
-            force_job(supercell, calculator)
+            static_job(supercell, **static_job_kwargs)
             for supercell in supercells
             if supercell is not None
         ]
 
     @job
-    def _thermo_job(
-        phonon: Phonopy, forces: list[NDArray], input_atoms
+    def _phonopy_thermo_job(
+        atoms: Atoms, force_job_results: list[dict]
     ) -> PhononSchema:
+        phonon = atoms_to_phonopy(atoms, supercell_matrix, atom_disp)
+        parameters = force_job_results[-1].get("parameters")
+        forces = [output["results"]["forces"] for output in force_job_results]
         phonon.forces = forces
         phonon.produce_force_constants()
         phonon.run_mesh()
@@ -88,15 +97,10 @@ def phonon_flow(
 
         return summarize_phonopy(
             phonon,
-            calculator,
-            input_atoms=input_atoms,
+            input_atoms=atoms,
+            parameters=parameters,
             additional_fields=fields_to_store,
         )
 
-    phonon = atoms_to_phonopy(atoms, supercell_matrix, atom_disp)
-    supercells = [
-        phonopy_atoms_to_ase_atoms(s) for s in phonon.supercells_with_displacements
-    ]
-    forces = _force_job_distributed(supercells)
-
-    return _thermo_job(phonon, forces, atoms)
+    force_job_results = _phonopy_forces_subflow(atoms)
+    return _phonopy_thermo_job(atoms, force_job_results)