+++ src/quacc/recipes/emt/core.py
@@ -20,7 +20,7 @@ if TYPE_CHECKING:
 
     from ase import Atoms
 
-    from quacc.schemas.ase import OptSchema, RunSchema
+    from quacc.schemas._aliases.ase import OptSchema, RunSchema
 
 
 @job