+++ src/quacc/recipes/vasp/core.py
@@ -14,7 +14,7 @@ if TYPE_CHECKING:
 
     from ase import Atoms
 
-    from quacc.schemas.vasp import VaspSchema
+    from quacc.schemas._aliases.vasp import VaspSchema
 
     class DoubleRelaxSchema(VaspSchema):
         relax1: VaspSchema